+++ app/src/main/java/woowacourse/omok/domain/model/Direction.kt
@@ -1,4 +1,4 @@
-package omok.model
+package woowacourse.omok.domain.model
 
 enum class Direction(val row: Int, val col: Int) {
     UP(-1, 0),
