+++ app/src/main/java/woowacourse/omok/domain/model/Stone.kt
@@ -1,4 +1,4 @@
-package omok.model
+package woowacourse.omok.domain.model
 
 enum class Stone {
     BLACK,
